requirements.txt
streamlit
pandas
numpy
matplotlib
seaborn
wordcloud
gensim
scikit-learn
nltk
beautifulsoup4
requests
textblob
lxml
newsapi-python